+++ libs/checkpoint-postgres/langgraph/store/postgres/__init__.py
@@ -1,4 +1,4 @@
-from langgraph.store.postgres.aio_store import AsyncPostgresStore
-from langgraph.store.postgres.base_store import PostgresStore
+from langgraph.store.postgres.aio import AsyncPostgresStore
+from langgraph.store.postgres.base import PostgresStore
 
 __all__ = ["AsyncPostgresStore", "PostgresStore"]
